The requirement is to display the basic AO dashboard with  a label displaying the assigned organisation name.
I've added dfesign api call to get a list of user organisations, this adds the matching service organisation to the claims. The call happens at point of token validation. User can have 1 or many organisations assigned but uses only the 1 assigned by the service.

Updated tests on dfesigninserviceApi  to include organisationCall.

I've also updated the viewdata title for each of the dashboard pages so it's appropriate i.e. apply dashboard, review dashboard and admin dashboard.

Need to add service layer tests for dfesigninservice. It's currently uses automoq carried over from the dfesignin package, I've spoken to hamza re this and will create a separate task to resolve.
I did some minor amendments concerning the usings on this class  which is why i've included it in this commit, I've commented it out as I'm going to create a separate task as technical debt.